requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
anyio==3.7.1
attrs==23.1.0
bcrypt==4.1.2
black==23.12.0
branca==0.7.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.7
ecdsa==0.18.0
fastapi==0.104.1
folium==0.15.1
frozenlist==1.4.1
geographiclib==2.0
geopy==2.4.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
Jinja2==3.1.3
Mako==1.3.2
MarkupSafe==2.1.5
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.4
packaging==23.2
pandas==2.2.0
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.1.0
psycopg2-binary==2.9.9
pyasn1==0.5.1
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2024.1
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.26
starlette==0.27.0
typing_extensions==4.8.0
tzdata==2023.4
urllib3==2.2.0
uvicorn==0.24.0.post1
xyzservices==2023.10.1
yarl==1.9.4
